fix checking if we should remove port group
